requirements.txt
asgiref==3.8.1
Django==4.2.16
filelock==3.16.1
fsspec==2024.9.0
jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.1
pillow==10.4.0
sqlparse==0.5.1
sympy==1.13.3
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing-extensions==4.12.2
